refactor: 에러코드 각 도메인 패키지로 이동
