+++ src/main/java/com/cotato/itda/domain/diary/service/command/DiaryCommentCommandService.java
@@ -11,7 +11,7 @@ import com.cotato.itda.domain.diary.repository.DiaryRepository;
 import com.cotato.itda.domain.diary.validator.DiaryAccessValidator;
 import com.cotato.itda.domain.member.entity.Member;
 import com.cotato.itda.domain.member.repository.MemberRepository;
-import com.cotato.itda.global.error.constant.DiaryErrorCode;
+import com.cotato.itda.domain.diary.exception.code.DiaryErrorCode;
 import com.cotato.itda.global.error.constant.UserErrorCode;
 import com.cotato.itda.global.error.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
